refactor(seo): cleanup logic
