+++ src/pages/speakers.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 import Layout from "../components/layout";
-import SEO from "../components/seo";
+import SEO from "../components/SEO/seo";
 
 import Nav from "../components/nav";
 import Breadcrumb from "../components/breadcrumb";
