+++ src/pages/tailwind-config-demo.js
@@ -1,7 +1,7 @@
 import React from "react";
 
 import Layout from "../components/layout";
-import SEO from "../components/seo";
+import SEO from "../components/SEO/seo";
 
 function TailwindDemoPage() {
   return (
